requirements.txt
django
django-bootstrap3
django-emoji
django-rest-swagger
djangorestframework
jsonfield
